requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-WTF==1.1.1
SQLAlchemy==2.0.22
Flask-SQLAlchemy==3.0.3
WTForms==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7